Move selective_state_update_ref to new utils file
